feat: deprecate check_catalog, and replace it by identical boundary_check function (with correct scope) in snapshot products helpers
